requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
yfinance
alpaca-py
sqlalchemy
python-dotenv
matplotlib
openai
